replaced homepage image with gatsby image on test site
