+++ examples/test-site/src/data/pages/header_image.json
@@ -1,4 +1,5 @@
 {
   "src": "/images/bodilessheader.jpg",
-  "alt": "BodilessJS Header"
+  "alt": "BodilessJS Header",
+  "preset": "fluid"
 }
